+++ src/state/state_change.rs
@@ -14,7 +14,7 @@ pub enum StateChange {
     PlayStatus(PlayStatus),
     NextScreen(Screen),
     NextElement(Element),
-    Index(Output, u32),
+    Index(Output, usize),
     None,
 }
